Add SHA 256 computation code
There is a main in this file right now. I will move that logic out into the
main.c file when I get time. I believe this won't compile (shoot me for saying
that if you want). The defines are out of order in the file. Will be fixed with
the next commit.