Add BSD Tag support
This modifies the hash algorithm abstraction further, which will be modified
once more when hash checking is implemented. It's likely that the SHA-2 will
only have 2 different implementations of the output algorithms (one pair for
SHA-256 based algos, and the other pair for the SHA-512 based algos).

In addition, this change requires that the developer use the BSD Tag for the
algorithm name (which isn't a big deal, since the find_algo_by_name is case
insensitive).